Basic software limit for shoulder
